Add files via upload
Wrote this off the back of an introductory creative artificial intelligence course I did. It reads books sentence by sentence and for each word creates a frequency table of words that follow that word in a sentence. Using this as the basis of predictive analysis as well as some fun games based on response to prompt and random sentence generation.